refactor: break the short import cycles
Six two- and three-module cycles, each closed by one function or a few
constants living on the wrong side of an edge: escapeXml, the compartment
grid primitives, the poller's pull state, the print-settings constants,
the lid constants hingeLidPlan read through the shared facade, and the
api's sanitizeString. Each moves to a leaf module its previous home now
re-exports, so no import path changes for callers.
