+++ src/core/sync/session/useSession.test.ts
@@ -132,7 +132,7 @@ describe('useSessionLifecycle', () => {
     // sign-in that picks 'merge' in the mismatch dialog, leaking A's
     // edits into B's account.
     const outboxModule = await import('../outbox');
-    const pollerModule = await import('../poller');
+    const pollerModule = await import('../pullState');
     const clearAllSpy = vi.spyOn(outboxModule, 'clearAll').mockResolvedValueOnce();
     const resetPullStateSpy = vi.spyOn(pollerModule, 'resetPullState').mockImplementation(() => {});
